core/requirements.txt
openai
pyyaml
python-dotenv
tabulate
